Added HookSignal to v2
