+++ kpi/serializers/v1/asset.py
@@ -4,6 +4,7 @@ from __future__ import absolute_import
 from kpi.serializers.v2.asset import AssetSerializer as AssetSerializerV2, \
     AssetListSerializer as AssetListSerializerV2, \
     AssetUrlListSerializer as AssetUrlListSerializerV2
+
 from .object_permission import ObjectPermissionNestedSerializer
