+++ kpi/urls/router_api_v2.py
@@ -6,7 +6,6 @@ from rest_framework_extensions.routers import ExtendedDefaultRouter
 from kobo.apps.hook.views import HookViewSet, HookLogViewSet
 from kpi.views.v1 import (
     AssetFileViewSet,
-    HookSignalViewSet,
     SubmissionViewSet,
 )
 
@@ -14,6 +13,7 @@ from kpi.views.v2.asset import AssetViewSet
 from kpi.views.v2.asset_snapshot import AssetSnapshotViewSet
 from kpi.views.v2.asset_version import AssetVersionViewSet
 from kpi.views.v2.collection import CollectionViewSet
+from kpi.views.v2.hook_signal import HookSignalViewSet
 from kpi.views.v2.user import UserViewSet
