+++ kpi/views/v1/hook_signal.py
@@ -1,20 +1,16 @@
 # -*- coding: utf-8 -*-
 from __future__ import unicode_literals, absolute_import
 
-from django.shortcuts import get_object_or_404
-from django.utils.translation import ugettext_lazy as _
-from rest_framework import status, viewsets
-from rest_framework.response import Response
-from rest_framework_extensions.mixins import NestedViewSetMixin
-
-from kobo.apps.hook.utils import HookUtils
-from kpi.models import Asset
-from kpi.utils.log import logging
+from kpi.views.v2.hook_signal import HookSignalViewSet as HookSignalViewSetV2
 
 
-class HookSignalViewSet(NestedViewSetMixin, viewsets.ViewSet):
+class HookSignalViewSet(HookSignalViewSetV2):
     """
-    ##
+    ## This document is for a deprecated version of kpi's API.
+
+    **Please upgrade to latest release `/api/v2/collections/`**
+
+
     This endpoint is only used to trigger asset's hooks if any.
 
     Tells the hooks to post an instance to external servers.
@@ -35,46 +31,5 @@ class HookSignalViewSet(NestedViewSetMixin, viewsets.ViewSet):
     >        }
 
     """
-    parent_model = Asset
 
-    def create(self, request, *args, **kwargs):
-        """
-        It's only used to trigger hook services of the Asset (so far).
-
-        :param request:
-        :return:
-        """
-        # Follow Open Rosa responses by default
-        response_status_code = status.HTTP_202_ACCEPTED
-        response = {
-            "detail": _(
-                "We got and saved your data, but may not have fully processed it. You should not try to resubmit.")
-        }
-        try:
-            asset_uid = self.get_parents_query_dict().get("asset")
-            asset = get_object_or_404(self.parent_model, uid=asset_uid)
-            instance_id = request.data.get("instance_id")
-            instance = asset.deployment.get_submission(instance_id)
-
-            # Check if instance really belongs to Asset.
-            if not (instance and instance.get(asset.deployment.INSTANCE_ID_FIELDNAME) == instance_id):
-                response_status_code = status.HTTP_404_NOT_FOUND
-                response = {
-                    "detail": _("Resource not found")
-                }
-
-            elif not HookUtils.call_services(asset, instance_id):
-                response_status_code = status.HTTP_409_CONFLICT
-                response = {
-                    "detail": _(
-                        "Your data for instance {} has been already submitted.".format(instance_id))
-                }
-
-        except Exception as e:
-            logging.error("HookSignalViewSet.create - {}".format(str(e)))
-            response = {
-                "detail": _("An error has occurred when calling the external service. Please retry later.")
-            }
-            response_status_code = status.HTTP_500_INTERNAL_SERVER_ERROR
-
-        return Response(response, status=response_status_code)
+    URL_NAMESPACE = None
